feat: faucet error
