+++ projects/main/src/app/models/faucets/faucet.infrastructure.service.ts
@@ -3,8 +3,6 @@ import { InterfaceFaucetInfrastructureService } from './faucet.service';
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { ConfigService } from 'projects/main/src/app/models/config.service';
-import { Observable, of } from 'rxjs';
-import { map } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root',
@@ -12,21 +10,18 @@ import { map } from 'rxjs/operators';
 export class FaucetInfrastructureService implements InterfaceFaucetInfrastructureService {
   constructor(private configS: ConfigService, private http: HttpClient) {}
 
-  postFaucetRequest$(faucetRequest: FaucetRequest): Observable<FaucetResponse> {
+  async postFaucetRequest(faucetRequest: FaucetRequest): Promise<FaucetResponse> {
     const requestBody = {
       address: faucetRequest.address,
       coins: faucetRequest.coins.map((coin) => coin.amount + coin.denom),
     };
     const faucetURL = this.getFaucetURL(faucetRequest.coins[0].denom);
     if (faucetURL !== undefined) {
-      return this.http
-        .post<FaucetResponse>(faucetURL, requestBody)
-        .pipe(map((response) => response));
+      return this.http.post<FaucetResponse>(faucetURL, requestBody).toPromise();
     } else {
-      return of({
-        address: faucetRequest.address,
+      return {
         transfers: [],
-      });
+      };
     }
   }
