+++ projects/main/src/app/models/faucets/faucet.service.ts
@@ -1,10 +1,10 @@
-import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
 import { FaucetInfrastructureService } from './faucet.infrastructure.service';
 import { FaucetRequest, FaucetResponse } from './faucet.model';
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
 
 export interface InterfaceFaucetInfrastructureService {
-  postFaucetRequest$: (faucetRequest: FaucetRequest) => void;
+  postFaucetRequest: (faucetRequest: FaucetRequest) => void;
 }
 
 @Injectable({
@@ -13,7 +13,7 @@ export interface InterfaceFaucetInfrastructureService {
 export class FaucetService {
   constructor(private faucetInfrastructureService: FaucetInfrastructureService) {}
 
-  postFaucetRequest$(faucetRequest: FaucetRequest): Observable<FaucetResponse> {
-    return this.faucetInfrastructureService.postFaucetRequest$(faucetRequest);
+  postFaucetRequest(faucetRequest: FaucetRequest): Promise<FaucetResponse> {
+    return this.faucetInfrastructureService.postFaucetRequest(faucetRequest);
   }
 }
